Frontend/classroom video player

+++ frontEnd/src/Components/ui/CommentCard/CommentCard.tsx
@@ -8,6 +8,7 @@ import user from '../../../redux/modules/user/user';
 import NewSubComment from '../Comments/NewCommentForm/NewSubComment';
 import SubCommentCard from '../SubCommentCard/SubCommentCard';
 import CommentCardHeader from './CommentCardHeader';
+import styles from './styles.module.scss';
 
 interface commentCardProp {
   comment: ICommentsWithUserData
@@ -17,13 +18,17 @@ export default function CommentCard({comment}: commentCardProp) {
   const [showSubComments, setShowSubComments] = useState(false);
   const [editedValue, setEditedValue] = useState('');
   const [edit, setEdit] = useState(false);
+  const [confirmDelete, setConfirmDelete] = useState(false);
 
   const { userData } = useSelector(({ user }: globalState) => user);
 
   const dispatch = useDispatch();
 
   const deleteComment = () => {
+    if(edit) return setEdit(!edit);
+    if(!confirmDelete) return setConfirmDelete(true);
     dispatch(deleteCommentAction({id: comment.id, userId: Number(userData.id)}, comment));
+    setConfirmDelete(false);
   };
 
   const handleEdit = () => {
@@ -37,31 +42,41 @@ export default function CommentCard({comment}: commentCardProp) {
   };
 
   return (
-    <section>
-      <CommentCardHeader userData={comment.userData} />
+    <section className={styles.card_caontainer}>
+      <CommentCardHeader userData={comment.userData}  />
       <article>
         {edit ? <input
           onChange={({target}) => setEditedValue(target.value)}
           value={editedValue}
-        /> :comment.content}
+        /> : <span className={styles.comment}>{comment.content}</span>}
         { comment.userData.id ===  userData.id &&
-        <button onClick={handleEdit}>
-          { edit ? 'Salvar' : 'Editar' }
-        </button>}
-        {
-          comment.userData.id === userData.id &&
-        <button onClick={deleteComment}>
-            deletar
-        </button>
+        <section>
+          <button onClick={handleEdit}>
+            { edit ? 'Salvar' : 'Editar' }
+          </button>
+          <button onClick={deleteComment}>
+            { confirmDelete ? 'Confirmar' : ( edit ? 'Cancelar' : 'deletar') }
+          </button>
+        </section>
         }
       </article>
-      <NewSubComment commentData={comment}/>
       {
-        showSubComments && comment.subComments.map((currSubComment: IsubComments) =>
-          <SubCommentCard subComment={currSubComment} key={currSubComment.id}/>)
+        showSubComments && comment.subComments.map((currSubComment: IsubComments, index: number) => {
+          return (
+            <section key={index}>
+              <SubCommentCard subComment={currSubComment} key={currSubComment.id}/>
+              {
+                index + 1 === comment.subComments.length && <NewSubComment commentData={comment}/>
+              }
+            </section>
+          );
+        })
       }
-      <button onClick={() => setShowSubComments(!showSubComments)}>
-        { showSubComments ? 'Fechar respostas' : 'Mostrar respostas' }
+      {
+        !comment.subComments.length && showSubComments && <NewSubComment commentData={comment}/>
+      }
+      <button className={styles.show_sub_comment} onClick={() => setShowSubComments(!showSubComments)}>
+        { showSubComments ? 'Esconder comentários' : 'Mostrar respostas'}
       </button>
     </section>
   );
